We have done a new repository from the previous one because the code and everything was getting too messy.
Now we think we are at the point that the main frame of our work is defined, so we think it is safe to start
the repository from this point and go on from here.

We divided our folder in: \data   where we do store the txt and pts of our graphs
                    \graph_maker  with the code that builds those files
                    \include      where we will put the classes and function to define and solve our problems
                    \src          where we will analyze our specific problems and test
                    \Prove        a folder where we can try out the code as we are building it and where
                                  we can be as messy as we want

The folder Prove will probably be deleted towards the and and probably more folder will be added as our
work goes on.

To understand how every code is working and how it should be run in every folder you can find a README.txt that
will, I hope, explain everything.